Refactor cudaq integration: separate conversion, annotation, and task logic
- cudaq_integration.py: focused on cudaq detection and cudaq→QIR→squin
  conversion (is_cudaq_kernel, cudaq_to_squin with run_pass included)
- logical_mvp.py: moved append_measurements_and_annotations here as
  general-purpose squin IR mutation (not cudaq-specific)
- device.py: task() now composes the two cleanly
- Remove debug print from analysis.py
